% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{calls}
\alias{calls}
\alias{ft_search}
\title{Calls}
\usage{
ft_search(q, ..., pages = 1, start_at_page = 1)
}
\arguments{
\item{q}{The search query.}

\item{...}{Any other parameter.}

\item{pages}{Number of pages to collect from \code{start_at_page}.}

\item{start_at_page}{Initial page.}
}
\description{
The content endpoint returns all pieces of content in the API.
}
\examples{
\dontrun{
(to_search <- gd_search("debates", pages = 13))
results <- gd_call(to_search)

# select items to retrieve
items_to_get <- gd_items(results$apiUrl[1:13])
items <- gd_call(items_to_get)
}

}
