% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{discover}
\alias{discover}
\alias{ft_list_aspects}
\alias{ft_list_curations}
\alias{ft_list_facets}
\alias{ft_list_searchable}
\alias{ft_list_sortable}
\title{Discover Methods}
\usage{
ft_list_aspects()

ft_list_curations()

ft_list_facets()

ft_list_searchable()

ft_list_sortable()
}
\description{
Useful lists to further filter \code{\link{ft_search}}.
}
\examples{
\dontrun{
ft_list_aspects()  
}

}
