% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_export.R
\name{Model}
\alias{Model}
\title{Create Model Object from the ornate::Model C++ Class}
\value{
A `Model` object from the _C++_ ornate::Model Class.
}
\description{
Allows for the creation of a Model Object in _C++_ from _R_
using the _C++_ ornate::Model class.
}
\examples{
##################
## Constructor

# Construct new FT::Model object called "m"
m = new(FT::Model)

##################
## Getters

m$train()

m$fit()

}
