% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pvalue_feature_select}
\alias{pvalue_feature_select}
\title{pvalue_feature_select}
\usage{
pvalue_feature_select(pvalues, threshold, has_intercept)
}
\arguments{
\item{pvalues}{pvalues}

\item{threshold}{threshold}

\item{has_intercept}{has_intercept}
}
\value{
selected feature idx
}
\description{
pvalue_feature_select
}
