% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\alias{ft_key}
\title{Setup}
\usage{
ft_key(key)
}
\arguments{
\item{key}{Your API key, freely available at \url{https://developer.ft.com}.}
}
\description{
Setup your session, all subsequent calls will be done using the API key.
}
\note{
You can specify \code{FT_API_KEY} as environment variable, likely in your \code{.Renviron} file.
}
\examples{
\dontrun{
guardian_key("xXXxxXxXxXXx")  
}

}
