% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{calls}
\alias{calls}
\alias{ft_search}
\title{Calls}
\usage{
ft_search(q, ..., pages = 1, start_at_page = 1)
}
\arguments{
\item{pages}{Number of pages to collect.}

\item{since}{The start date and time: any content created, updated or deleted since this date should be returned. 
A valid since parameter must be supplied, and it must be either a \code{Date} or \code{POSIX} object. 
The date and time must not be in the future, and cannot be older than 90 days from the present.}
}
\description{
The content endpoint returns all pieces of content in the API.
}
\examples{
\dontrun{
(to_search <- gd_search("debates", pages = 13))
results <- gd_call(to_search)

# select items to retrieve
items_to_get <- gd_items(results$apiUrl[1:13])
items <- gd_call(items_to_get)
}

}
